Android apk-based Gtest: Keep user data dir for PRE_ tests.
https://crrev.com/c/6599562 refactored the code to support the PRE_
test grouping in gtest. This CL is a followup to keep user data dir
for PRE_ tests in apk-based gtest.

**Note** that PRE_ tests in android_browsertests still don't work
because it uses a different way to setup and clean user data dir,
after the CL https://crrev.com/c/1677591

This CL introduces new intent extra "KeepUserDataDir" in
1. NativeTestInstrumentationTestRunner.
2. NativeTestIntent.

The python code local_device_gtest_run.py sets the extra #1 to true
for a PRE_ test group.

Then NativeTestInstrumentationTestRunner.java consumes this info and
sets the extra #2 for each individual java test shard.

The *BrowserTestActivity.java then read this info, and keep the user
data dir when needed, before running the tests.

This CL also enables NetworkServiceBrowserCacheResetTest in
content_browsertests on android, which contains PRE_ tests

Bug: 40200835
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6568919
Commit-Queue: Haiyang Pan <hypan@google.com>
Cr-Commit-Position: refs/heads/main@{#1473200}

+++ chrome/test/android/browsertests_apk/src/org/chromium/android_browsertests_apk/ChromeBrowserTestsActivity.java
@@ -37,7 +37,12 @@ public class ChromeBrowserTestsActivity extends ChromeTabbedActivity {
         mTest.appendCommandLineFlags(
                 "--remote-debugging-socket-name android_browsertests_devtools_remote");
 
-        NativeBrowserTest.deletePrivateDataDirectory(getPrivateDataDirectory());
+        // TODO(crbug.com/40200835): Enable PRE_ tests in android_browsertests
+        // Note that android_browsertests does not use this data directory. It
+        // uses a temp dir created by c++ code since https://crrev.com/c/1677591
+        if (!mTest.shouldKeepUserDataDir()) {
+            NativeBrowserTest.deletePrivateDataDirectory(getPrivateDataDirectory());
+        }
 
         // Replace ContentMain() with running our NativeTest suite.
         BrowserStartupController.getInstance()
