+++ testing/android/native_test/java/src/org/chromium/native_test/NativeBrowserTestActivity.java
@@ -12,7 +12,7 @@ import java.io.File;
 
 /** An {@link android.app.Activity} for running native browser tests. */
 public abstract class NativeBrowserTestActivity extends FragmentActivity {
-    private static final String TAG = "NativeTest";
+    private static final String TAG = "NativeTestActivity";
 
     private final NativeTest mTest = new NativeTest();
     private boolean mStarted;
@@ -44,7 +44,9 @@ public abstract class NativeBrowserTestActivity extends FragmentActivity {
         if (mStarted) return;
 
         mStarted = true;
-        NativeBrowserTest.deletePrivateDataDirectory(getPrivateDataDirectory());
+        if (!mTest.shouldKeepUserDataDir()) {
+            NativeBrowserTest.deletePrivateDataDirectory(getPrivateDataDirectory());
+        }
         initializeBrowserProcess();
     }
